+++ packages/vue-components/.storybook/preview.ts
@@ -4,13 +4,29 @@ import "vuetify/styles"
 import * as components from "vuetify/components"
 import * as directives from "vuetify/directives"
 
+import { aliases, mdi } from "vuetify/iconsets/mdi-svg"
+
+// Import highlight.js
+import VueHighlightJS from "vue3-highlightjs"
+import "highlight.js/styles/default.css" // Or your preferred theme
+
 const vuetify = createVuetify({
   components,
   directives,
+  icons: {
+    defaultSet: "mdi",
+    aliases,
+    sets: {
+      mdi,
+    },
+  },
 })
 
 setup(app => {
+  // Register Vuetify
   app.use(vuetify)
+  // Register highlight.js
+  app.use(VueHighlightJS)
 })
 
 const preview: Preview = {