+++ packages/vue-components/src/stories/OmegaForm.stories.ts
@@ -10,6 +10,10 @@ import SimpleFormVuetifyDefault from "./OmegaForm/SimpleFormVuetifyDefault.vue"
 import SumExample from "./OmegaForm/SumExample.vue"
 import PersistencyForm from "./OmegaForm/PersistencyForm.vue"
 import AutoGeneration from "./OmegaForm/AutoGeneration.vue"
+import MetaForm from "./OmegaForm/Meta.vue"
+import FormInput from "./OmegaForm/form.Input.vue"
+import OneHundredWaysToWriteAForm from "./OmegaForm/OneHundredWaysToWriteAForm.vue"
+import Clearable from "./OmegaForm/Clearable.vue"
 
 const mockIntl = {
   locale: ref("en"),
@@ -20,7 +24,6 @@ const mockIntl = {
 const meta: Meta<typeof OmegaForm> = {
   title: "Components/OmegaForm",
   component: OmegaForm,
-  tags: ["autodocs"],
   argTypes: {
     schema: { control: "object" },
     onSubmit: { action: "submitted" },
@@ -40,6 +43,13 @@ const meta: Meta<typeof OmegaForm> = {
 
 export default meta
 type Story = StoryObj<typeof meta>
+
+export const AnHundredWayToWriteAFormStory: Story = {
+  render: () => ({
+    components: { OneHundredWaysToWriteAForm },
+    template: "<OneHundredWaysToWriteAForm />",
+  }),
+}
 
 export const SimpleFormStory: Story = {
   render: () => ({
@@ -87,5 +97,26 @@ export const AutoGenerationStory: Story = {
   render: () => ({
     components: { AutoGeneration },
     template: "<AutoGeneration />",
+  }),
+}
+
+export const MetaStory: Story = {
+  render: () => ({
+    components: { MetaForm },
+    template: "<MetaForm />",
+  }),
+}
+
+export const FormInputStory: Story = {
+  render: () => ({
+    components: { FormInput },
+    template: "<FormInput />",
+  }),
+}
+
+export const ClearableStory: Story = {
+  render: () => ({
+    components: { Clearable },
+    template: "<Clearable />",
   }),
 }